Fix remote exec flaw
Since the description is provided by the user and unfiltered, someone
can provides a description with %(some shell command) that would be
executed by the web server.

While a solution would be to forbid the % caracter in description,
some people might want to use it, so better make sure it is properly
escaped by when feed to rpmbuild.