Match coprnames URL even without any user name
As the COPR is not widely used for building the collections these days,
the common practice was to submit the new collection form
without COPR user name.
Before this change, this would cause the respective URL pattern to not
match and attempted to render the request as a static page
– which would result in missing template exception/error.

After this change, even if no COPR user name is provided,
the URL still routes to the `coprnames` view.
This results in empty JSON list being returned and no more
confusing template errors.